+++ package.json
@@ -16,6 +16,7 @@
     "react-router": "^3.0.2",
     "react-tap-event-plugin": "^2.0.1",
     "redux": "^3.6.0",
+    "redux-form": "^6.5.0",
     "redux-thunk": "^2.2.0",
     "waves.js": "0.0.7-1"
   },
