+++ src/components/MainLayout/mainLayout.css
@@ -1,0 +1,4 @@
+.active {
+  background-color: rgba(0, 0, 0, 0.04) !important;
+  color: #17579c !important;
+}
