+++ src/components/Send/sendWizard.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+class SendWizard extends React.Component {
+
+  render() {
+    return (<div>Send wizard</div>);
+  }
+}
+
+export default SendWizard;
