+++ src/router.js
@@ -2,12 +2,13 @@ import React from 'react';
 import {Router, Route, browserHistory, IndexRoute} from 'react-router';
 
 import App from './app';
-import MainLayout from './components/main-layout';
+import MainLayout from './components/MainLayout';
 import Dashboard from './components/Dashboard';
 import Login from './components/Login';
 import EnsureLoggedInContainer from './components/ensure-logged-in-container';
 import Wallet from './components/Wallet';
 import Account from './components/Account';
+import SendWizard from './components/Send';
 
 const router = (
   <Router history={browserHistory}>
@@ -18,8 +19,11 @@ const router = (
         <Route component={MainLayout}>
           <IndexRoute component={Dashboard}/>
           <Route path="dashboard" component={Dashboard}/>
-          <Route path="wallet" component={Wallet} />
-          <Route path="account/:address" component={Account} />
+          <Route path="wallet">
+            <IndexRoute component={Wallet}/>
+            <Route path="account/:address/send" component={SendWizard} />
+            <Route path="account/:address" component={Account} />
+          </Route>
         </Route>
       </Route>
     </Route>
